Moved BackupContainerFileSystem implementation to cpp file
